+++ engine/.babelrc
@@ -1,4 +1,14 @@
 {
-  "presets": ["@babel/preset-env"],
-  "exclude": ["node_modules/**"]
+  "presets": [
+    ["@babel/preset-env",
+    {
+      "loose": true,
+      "targets": {
+        "node": "12"
+      }
+    }]
+  ],
+  "exclude": ["node_modules/**"],
+  "plugins": [ "@babel/plugin-proposal-export-default-from" ]
 }
+
