+++ engine/package.json
@@ -5,6 +5,8 @@
   "main": "index.js",
   "scripts": {
     "lint": "standard",
+    "build": "babel src -d lib",
+    "build:watch": "babel src -d lib --watch",
     "test": "lab -T node_modules/@xgovformbuilder/lab-babel -v -P 'test' test/cases -r console -o stdout -r html -o unit-test.html -I version"
   },
   "repository": {
@@ -25,7 +27,9 @@
     "wreck": "14.2.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.10.1",
     "@babel/core": "^7.10.2",
+    "@babel/plugin-proposal-export-default-from": "^7.10.1",
     "@babel/preset-env": "^7.10.2",
     "@hapi/code": "^8.0.1",
     "@hapi/lab": "^22.0.4",
