+++ engine/src/lib.js
@@ -1,0 +1,11 @@
+import findByPostcode from './address-service'
+import * as ComponentTypes from './component-types'
+import * as ConditionalComponentTypes from './conditional-component-types'
+import * as Helpers from './helpers'
+import Model from './model'
+import Page  from './page'
+import Schema from './schema'
+import { plugin } from './index'
+
+export { findByPostcode, ComponentTypes, ConditionalComponentTypes, Helpers, Model, Page, Schema }
+export default plugin
