+++ runner/src/server/plugins/builder/index.js
@@ -5,6 +5,7 @@ const config = require('../../config')
 
 const relativeTo = __dirname
 const defaultPageController = './pages'
+import plugin from 'digital-form-builder-engine'
 
 const configFiles = fs.readdirSync(configPath).filter(filename => {
   if (filename.indexOf('.json') >= 0) {
@@ -36,7 +37,7 @@ const configurePlugins = (configFile, customPath) => {
   }
 
   return {
-    plugin: require('digital-form-builder-engine'),
+    plugin,
     options: { modelOptions, configs, previewMode: config.previewMode }
   }
 }
