+++ runner/src/server/plugins/builder/model.js
@@ -1,5 +1,5 @@
 // const joi = require('joi')
-const EngineModel = require('digital-form-builder-engine/model')
+const { EngineModel } = require('digital-form-builder-engine')
 
 class Model extends EngineModel {
   makeSchema (state) {
